test: unify jest setup
