+++ tests/setup/integration.setup.js
@@ -1,12 +1,11 @@
+require('./setupTests.ts');
+
 // Integration test setup for repository and API tests
 // Node.js environment - includes MSW for network mocking
 //
 // Usage: Import this file in integration tests that need MSW:
 // setupFilesAfterEnv: ['<rootDir>/tests/setup/integration.setup.js']
 // or add this comment to test files: @jest-environment-options {"setupFilesAfterEnv": ["<rootDir>/tests/setup/integration.setup.js"]}
-
-// Ensure fetch is available in Node.js environment for API tests
-require('cross-fetch/polyfill');
 
 // Import MSW server
 const { server } = require('./msw/server');
